Mapping of the MS Features back to LCMS data.

+++ src/MultiAlignCore/IO/Features/DataAccessFactory.cs
@@ -1,4 +1,5 @@
-﻿using System.IO;
+﻿using System;
+using System.IO;
 using MultiAlignCore.Data;
 using MultiAlignCore.IO.Features.Hibernate;
 
@@ -7,24 +8,40 @@ namespace MultiAlignCore.IO.Features
     public class DataAccessFactory
     {
 
-        public static FeatureDataAccessProviders CreateDataAccessProviders(MultiAlignAnalysis analysis)
+        public static FeatureDataAccessProviders CreateDataAccessProviders(MultiAlignAnalysis analysis, bool deleteIfExists)
         {
             string path = AnalysisPathUtils.BuildAnalysisName(analysis.MetaData.AnalysisPath, analysis.MetaData.AnalysisName);
+            return CreateDataAccessProviders(path, deleteIfExists);
+        }
 
-            bool exists = File.Exists(path);
-            if (exists)
+        public static FeatureDataAccessProviders CreateDataAccessProviders(string path, bool deleteIfExists)
+        {
+            if (deleteIfExists)
             {
-                File.Delete(path);
+                bool exists = File.Exists(path);
+                if (exists)
+                {
+                    File.Delete(path);
+                }
             }
-            NHibernateUtil.ConnectToDatabase(path, true);
+            try
+            {
+                NHibernateUtil.ConnectToDatabase(path, true);
+            }
+            catch(Exception ex)
+            {
+                throw ex;
+            }
             IUmcDAO featureCache            = new UmcDAOHibernate();
             IUmcClusterDAO clusterCache     = new UmcClusterDAOHibernate();
             IMSFeatureDAO msFeatureCache    = new MSFeatureDAOHibernate();
+            IGenericDAO<MSFeatureToLCMSFeatureMap> msFeatureMap = new GenericDAOHibernate<MSFeatureToLCMSFeatureMap>();
 
             FeatureDataAccessProviders providers =
                 new FeatureDataAccessProviders( featureCache, 
                                                 clusterCache,
-                                                msFeatureCache);
+                                                msFeatureCache,
+                                                msFeatureMap);
 
             return providers;
         }
